init share trade server - client - bigchaindb_common refactor
