--- bigchaindb_common/css/scss/sidebar.scss
@@ -1,62 +1,0 @@
-@import 'variables';
-
-
-#wrapper {
-  padding-left: $sidebar--wide;
-  transition: all 0.4s ease 0s;
-  margin-top: $navbar-height;
-}
-
-#sidebar-wrapper {
-  margin-left: -$sidebar--wide;
-  left: $sidebar--wide;
-  width: $sidebar--wide;
-  height: 100%;
-  overflow-y: auto;
-  z-index: 1200;
-  transition: all 0.4s ease 0s;
-}
-
-.sidebar-nav {
-  position: fixed;
-  top: $navbar-height;
-  width: $sidebar--wide;
-  margin: 0;
-  padding: 2em 0;
-  z-index: 1100;
-  height: 100vh;
-  border-right: 1px solid #888888;
-  ul {
-    list-style: none;
-    padding: 0;
-  }
-  .dropdown-menu {
-    overflow-y: scroll;
-    max-height: 50vh;
-  }
-
-}
-
-@media (max-width: 930px) {
-
-  #wrapper {
-    padding-left: 0;
-  }
-
-  #sidebar-wrapper {
-    left: 0;
-  }
-
-  #wrapper.active {
-    position: relative;
-    left: $sidebar--wide;
-  }
-
-  #wrapper.active #sidebar-wrapper {
-    left: $sidebar--wide;
-    width: $sidebar--wide;
-    transition: all 0.4s ease 0s;
-  }
-
-}
-
