--- bigchaindb_common/css/scss/style.scss
@@ -1,275 +1,0 @@
-@import "variables";
-
-/**
- * Colors and backgrounds
- */
-
-$row: 100%;
-$row--wide: 900px;
-$gutter: 20px;
-$small: 12px;
-
-$color--brand: #c90000;
-$color--background-secondary: #EEE;
-
-html {
-  font-family: 'Lato', sans-serif;
-}
-
-body {
-  background-color: $bg-color;
-}
-
-.App {
-  margin: auto;
-  //max-width: $content--max-width + $sidebar--wide;
-  height: 100%;
-}
-
-a {
-  color: $fg-color;
-}
-
-#page-content-wrapper {
-  width: 100%;
-}
-
-.page-content {
-  padding: 1em 1em 5em
-}
-.content-text{
-  margin-top: 2em;
-  text-align: center;
-}
-
-.hero,
-.page--result {
-  background-color: $color--background-secondary;
-}
-
-.logo-brand {
-  > span:first-of-type {
-    font-weight: 700;
-  }
-}
-
-.navbar-fixed-bottom {
-  margin-left: $sidebar--wide;
-}
-
-@media (max-width: 930px) {
-  .navbar-fixed-bottom {
-    margin-left: 0;
-  }
-}
-/**
- * Layout
- */
-
-.row {
-  max-width: $row;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.list-row,
-.list-item {
-  font-size: .9em;
-  text-transform: uppercase;
-  padding: 1em;
-  background-color: $bg-color;
-  border-bottom: 1px solid lighten($fg-color, 80%);
-  color: lighten($fg-color, 30%);
-}
-
-.list-row {
-  &:hover {
-    color: $bg-color !important;
-    .list-row-detail {
-      color: $bg-color !important;
-    }
-    background-color: lighten($bg-color--hover, 30%);
-    cursor: pointer;
-  }
-}
-
-.list-row.active {
-  color: $bg-color !important;
-  .list-row-detail {
-    color: $bg-color !important;
-  }
-  background-color: $bg-color--hover;
-}
-
-.list-row-name {
-  font-weight: bold;
-}
-
-.list-row-detail {
-  font-size: 0.8em;
-  font-style: italic;
-  color: lighten($fg-color, 50%);
-}
-
-.list-row-name,
-.list-row-detail {
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-.asset-container {
-  border: 1px solid lighten($fg-color, 50%);
-  background: lighten($fg-color, 95%);
-  padding: 1em;
-  border-radius: .7em;
-  margin-bottom: .5em;
-}
-
-.asset-container-id,
-.asset-container-timestamp {
-  font-size: 0.8em;
-  color: lighten($fg-color, 50%);
-  text-transform: uppercase;
-}
-
-.asset-container-id {
-  font-style: italic;
-  margin-bottom: 1em;
-  font-size: 0.75em;
-}
-
-.asset-container-timestamp {
-  margin-bottom: -.6em;
-}
-
-input {
-    width: 100%;
-    font-size: 20px;
-    line-height: 130%;
-    border: 1px solid #BBB;
-    padding: 1em;
-}
-
-
-.searchbar {
-  margin: 0 auto;
-}
-
-.searchbar input[type="url"] {
-  width: 80%;
-  display: inline-block;
-  @media (max-width: 480px) {
-    width: 70%;
-  }
-}
-
-.searchbar input[type="submit"] {
-  width: calc(20% - 15px);
-  padding-left: 0;
-  padding-right: 0;
-  margin-left: 15px;
-  margin-right: 0;
-  display: inline-block;
-  @media (max-width: 480px) {
-    width: 30%;
-    margin-left: 0px;
-  }
-}
-
-
-#app-container {
-  margin-top: 50vh;
-  text-align: center;
-  background: $fg-color;
-  color: black;
-}
-
-#container {
-  background-color: $color--background;
-}
-
-.content-header-wrapper {
-  background-color: $bg-color;
-  width: 100%;
-}
-
-.content-header {
-  width: 100%;
-  max-width: $content--max-width;
-  position: fixed;
-  background-color: $bg-color;
-  height: 70px;
-  box-shadow: 0 2px 2px -2px rgba(0, 0, 0, .5);
-  z-index: 1200;
-}
-
-.card {
-  max-width: $row--wide;
-  background: $bg-color;
-  box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.4);
-  padding: 20px 40px;
-  margin: 0 auto 40px;
-  overflow: hidden;
-
-  .desc {
-    font-size: 35px;
-    line-height: 150%;
-    margin: 0 0 25px 0;
-
-    @media (max-width: 600px) {
-      font-size: 20px;
-    }
-
-    strong {
-      background-color: black;
-      color: $fg-color;
-      padding: 4px 20px;
-    }
-  }
-}
-
-.card--summary {
-  .desc {
-    float: left;
-    width: 60%;
-    @media (max-width: 600px) {
-      float: none;
-      width: 100%;
-      display: block;
-      margin-left: auto;
-      margin-right: auto;
-    }
-  }
-
-  .preview {
-    margin-top: -10px;
-    float: right;
-    width: 30%;
-    max-width: 400px;
-    @media (max-width: 600px) {
-      float: none;
-      width: 100%;
-      margin-bottom: 30px;
-    }
-  }
-}
-
-.card {
-  .disclaimer {
-    padding-top: 30px;
-    text-align: right;
-    font-size: 12px;
-  }
-}
-
-.card--claim {
-  box-shadow: none;
-  background-color: transparent;
-  .desc {
-    font-size: 20px;
-    text-align: center;
-    a {
-      color: inherit;
-    }
-  }
-}
