--- bigchaindb_common/js/constants/application_constants.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const apiEndpoint = 'http://localhost:8000/api/';
-const baseUrl = '/';
-
-const constants = {
-    baseUrl,
-    apiEndpoint
-};
-
-export default constants;
