--- bigchaindb_common/js/react/alt.js
@@ -1,5 +1,0 @@
-'use strict';
-
-import Alt from 'alt';
-
-export let alt = new Alt();
