+++ client/share_trade/scss/main.scss
@@ -1,0 +1,15 @@
+@import url(https://fonts.googleapis.com/css?family=Lato:300);
+@import url(https://fonts.googleapis.com/css?family=Oswald);
+@import url(https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css);
+
+$icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
+@import "../../../node_modules/bootstrap-sass/assets/stylesheets/_bootstrap.scss";
+@import '../../lib/css/scss/normalize.css';
+@import '../../lib/css/scss/custom_bootstrap';
+@import '../../lib/css/scss/spinner';
+@import '../../lib/css/scss/sidebar';
+@import '../../lib/css/scss/style';
+@import './style';
+
+
+
