--- on_the_record/client/scss/main.scss
@@ -1,11 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Lato:300);
-@import url(https://fonts.googleapis.com/css?family=Oswald);
-@import url(https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css);
-
-$icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
-@import "../../../node_modules/bootstrap-sass/assets/stylesheets/_bootstrap.scss";
-@import '../../../bigchaindb_common/css/scss/normalize.css';
-@import '../../../bigchaindb_common/css/scss/custom_bootstrap';
-@import '../../../bigchaindb_common/css/scss/spinner';
-@import '../../../bigchaindb_common/css/scss/sidebar';
-@import '../../../bigchaindb_common/css/scss/style';
