--- on_the_record/server/templates/index.html
@@ -1,10 +1,0 @@
-{% extends "_base.html" %}
-
-
-{% block body %}
-
-<div id="container"></div>
-
-<!--<script type="text/javascript" src="../../client/bundles/"></script>-->
-<script type="text/javascript" src="{{ url_for('static', filename='app.js') }}"></script>
-{% endblock %}
