+++ package.json
@@ -47,6 +47,7 @@
     "react": "^0.14.2",
     "react-bootstrap": "^0.28.1",
     "react-dom": "0.14.0",
+    "react-matrix": "0.0.6",
     "react-router": ">=2.0.0-rc2",
     "react-router-bootstrap": "^0.20.0",
     "react-scroll": "^1.0.15",
