+++ server/templates/on_the_record.html
@@ -1,0 +1,10 @@
+{% extends "_base_on_the_record.html" %}
+
+
+{% block body %}
+
+<div id="container"></div>
+
+<!--<script type="text/javascript" src="../../client/bundles/"></script>-->
+<script type="text/javascript" src="{{ url_for('static', filename='on_the_record.js') }}"></script>
+{% endblock %}
