+++ server/templates/share_trade.html
@@ -1,0 +1,10 @@
+{% extends "_base_share_trade.html" %}
+
+
+{% block body %}
+
+<div id="container"></div>
+
+<!--<script type="text/javascript" src="../../client/bundles/"></script>-->
+<script type="text/javascript" src="{{ url_for('static', filename='share_trade.js') }}"></script>
+{% endblock %}
